Fix metadata calculation issues

+++ src/lib/metadata/lofar_udp_metadata_GUPPI.c
@@ -148,7 +148,7 @@ int lofar_udp_metadata_update_GUPPI(lofar_udp_metadata *metadata, int newObs) {
 		return -1;
 	}
 
-	metadata->output.guppi->dropblk = metadata->upm_last_dropped_packets / metadata->upm_pack_per_iter / metadata->upm_num_inputs;
+	metadata->output.guppi->dropblk = (float) metadata->upm_last_dropped_packets / (float) metadata->upm_pack_per_iter / (float) metadata->upm_num_inputs;
 	metadata->output.guppi->droptot = metadata->upm_dropped_packets / metadata->upm_processed_packets;
 	metadata->output.guppi->pktidx = (metadata->upm_processed_packets / metadata->upm_num_inputs) - metadata->upm_pack_per_iter;
